account-service/requirements.txt
flask
flask_login
flask_wtf
wtforms
wtforms.validators
flask_bcrypt
psycopg2-binary
requests